requirements.txt
flask>=3.0.0
line-bot-sdk>=3.11.0
groq>=0.9.0
requests>=2.32.0
